misc/HtmlTestSuite/results/test-0543.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml">
  <head>
    <link href="https://drafts.csswg.org/css-env-1/" rel="help"/>
    <title>Test that nested var() fallback values work with CSS env vars</title>
    <script src="/resources/testharness.js"/>
    <script src="/resources/testharnessreport.js"/>
    <style>
      body {
        --main-bg-color: rgb(0, 128, 0);
        background-color: env(test, var(--main-bg-color));
      }
    </style>
  </head>
  <body>
    <script>
    test(() =&gt; {
      const style = window.getComputedStyle(document.body);
      assert_equals(style.getPropertyValue("background-color"), "rgb(0, 128, 0)");
    });
    </script>
  </body>
</html>
